Refactor Claude client implementation by moving main functionality to index.ts
- Deleted the client.ts file and migrated its content to index.ts for better organization.
- Added ANSI color coding for improved console output during API requests.
- Implemented non-streaming and streaming examples with enhanced logging and statistics.
- Updated package.json to reflect the new entry point for the start script.
